# packages/api/tsoa-build/swagger-spec.json
{
	"components": {
...
		"schemas": {
			"TutorId": {
				"type": "string",
				"description": "A tutor user id"
			},
			"StudentId": {
				"type": "string",
				"description": "A student user id"
			},
			"UserId": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/TutorId"
					},
					{
						"$ref": "#/components/schemas/StudentId"
					}
				],
				"description": "A general user id, ie a student or tutor id"
			},
			"UrlString": {
				"type": "string",
				"description": "An internet link"
			},
			"Country": {
				"type": "string",
				"enum": [
					"American Samoa",
					"Anguilla",
					"Antigua and Barbuda",
					"Australia",
					"Austria",
					"Bahamas",
					"Barbados",
					"Belgium",
					"Belize",
					"Bermuda",
					"Botswana",
					"British Indian Ocean Territory",
					"British Virgin Islands",
					"Burundi",
					"Cameroon",
					"Canada",
					"Caribbean",
					"Cayman Islands",
					"Christmas Island",
					"Cocos [Keeling] Islands",
					"Cook Islands",
					"Cyprus",
					"Denmark",
					"Dominica",
					"Eritrea",
					"Europe",
					"Falkland Islands",
					"Finland",
					"Fiji",
					"Gambia",
					"Germany",
					"Ghana",
					"Gibraltar",
					"Grenada",
					"Guam",
					"Guernsey",
					"Guyana",
					"Hong Kong",
					"India",
					"Ireland",
					"Isle of Man",
					"Israel",
					"Jamaica",
					"Jersey",
					"Kenya",
					"Kiribati",
					"Lesotho",
					"Liberia",
					"Macao SAR",
					"Madagascar",
					"Malawi",
					"Malaysia",
					"Malta",
					"Marshall Islands",
					"Mauritius",
					"Micronesia",
					"Montserrat",
					"Namibia",
					"Nauru",
					"Netherlands",
					"New Zealand",
					"Nigeria",
					"Niue",
					"Norfolk Island",
					"Northern Mariana Islands",
					"Pakistan",
					"Palau",
					"Papua New Guinea",
					"Pitcairn Islands",
					"Puerto Rico",
					"Republic of the Philippines",
					"Rwanda",
					"Saint Kitts and Nevis",
					"Saint Lucia",
					"Saint Vincent and the Grenadines",
					"Samoa",
					"Seychelles",
					"Sierra Leone",
					"Singapore",
					"Sint Maarten",
					"Slovenia",
					"Solomon Islands",
					"South Africa",
					"South Sudan",
					"St Helena, Ascension,  Tristan da Cunha",
					"Sudan",
					"Swaziland",
					"Sweden",
					"Switzerland",
					"Tanzania",
					"Tokelau",
					"Tonga",
					"Trinidad and Tobago",
					"Turks and Caicos Islands",
					"Tuvalu",
					"Uganda",
					"United Arab Emirates",
					"United Kingdom",
					"United States",
					"US Minor Outlying Islands",
					"US Virgin Islands",
					"Vanuatu",
					"World",
					"Zambia",
					"Zimbabwe",
					"Algeria",
					"Benin",
					"Burkina Faso",
					"Central African Republic",
					"Chad",
					"Comoros",
					"Congo",
					"Congo, DRC",
					"Côte d'Ivoire",
					"Djibouti",
					"Equatorial Guinea",
					"France",
					"French Guiana",
					"French Polynesia",
					"Gabon",
					"Guadeloupe",
					"Guinea",
					"Haiti",
					"Luxembourg",
					"Mali",
					"Martinique",
					"Mauritania",
					"Mayotte",
					"Morocco",
					"New Caledonia",
					"Niger",
					"Principality of Monaco",
					"Reunion",
					"Saint Barthélemy",
					"Saint Martin",
					"Saint Pierre and Miquelon",
					"Senegal",
					"Syria",
					"Togo",
					"Tunisia",
					"Wallis and Futuna",
					"Brunei Darussalam",
					"Angola",
					"Brazil",
					"Cabo Verde",
					"Guinea-Bissau",
					"Mozambique",
					"Portugal",
					"São Tomé and Príncipe",
					"Timor-Leste"
				],
				"description": "A country name e.g. \"Malaysia\""
			},
			"LocaleCode": {
				"type": "string",
				"enum": [
					"en",
					"fr",
					"pt",
					"ms"
				],
				"description": "A locale code to specify language e.g. en for English"
			},
			"PublicStudentData": {
				"properties": {
					"id": {
						"$ref": "#/components/schemas/UserId",
						"description": "Unique id for a user"
					},
					"imageUrl": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UrlString"
							}
						],
						"nullable": true,
						"description": "Link to profile image, if defined"
					},
					"introduction": {
						"type": "string",
						"description": "Optional text for users to introduce themselves,\r\ndescribe their experience, what subjects they are interested in etc"
					},
					"prefferedCountries": {
						"items": {
							"$ref": "#/components/schemas/Country"
						},
						"type": "array",
						"description": "The countries the user prefers for content"
					},
					"prefferedLocales": {
						"items": {
							"$ref": "#/components/schemas/LocaleCode"
						},
						"type": "array",
						"description": "The locales a user prefers for content"
					},
					"userName": {
						"type": "string",
						"description": "Username which will be displayed for a user"
					}
				},
				"required": [
					"id",
					"prefferedCountries",
					"prefferedLocales",
					"userName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SubjectId": {
				"type": "string",
				"description": "A general subject id that doesnt specify locale or country"
			},
...
			"ConfidenceLevelEnum": {
				"description": "Confidence level options mapped to integers",
				"enum": [
					0,
					1,
					2,
					3
				],
				"type": "number"
			},
			"UserSubjectData": {
				"description": "Data about a subject specific to a user, as it is stored in database\r\n// ? should this be public?",
				"properties": {
					"id": {
						"$ref": "#/components/schemas/SubjectId",
						"description": "Unique id to the related subject"
					},
					"confidenceLevel": {
						"$ref": "#/components/schemas/ConfidenceLevelEnum",
						"description": "A rating of how confident a user feels in the subject, the numbers"
					},
					"country": {
						"$ref": "#/components/schemas/Country"
					},
					"detail": {
						"type": "string",
						"description": "Detail of the users experience with the subject"
					},
					"locale": {
						"$ref": "#/components/schemas/LocaleCode"
					}
				},
				"required": [
					"id",
					"confidenceLevel",
					"country",
					"detail",
					"locale"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"EmailString": {
				"type": "string",
				"description": "A valid email address"
			},
...
			"LinkedTutorData": {
				"description": "Data about a tutor from the perspective of a student, as it is stored in database",
				"properties": {
					"id": {
						"$ref": "#/components/schemas/TutorId",
						"description": "Id of the tutor related to the student"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PrivateStudentData": {
				"description": "Schema of data that represents a student, as it is stored in database",
				"properties": {
					"id": {
						"$ref": "#/components/schemas/UserId",
						"description": "Unique id for a user"
					},
					"imageUrl": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UrlString"
							}
						],
						"nullable": true,
						"description": "Link to profile image, if defined"
					},
					"introduction": {
						"type": "string",
						"description": "Optional text for users to introduce themselves,\r\ndescribe their experience, what subjects they are interested in etc"
					},
					"prefferedCountries": {
						"items": {
							"$ref": "#/components/schemas/Country"
						},
						"type": "array",
						"description": "The countries the user prefers for content"
					},
					"prefferedLocales": {
						"items": {
							"$ref": "#/components/schemas/LocaleCode"
						},
						"type": "array",
						"description": "The locales a user prefers for content"
					},
					"userName": {
						"type": "string",
						"description": "Username which will be displayed for a user"
					},
					"email": {
						"$ref": "#/components/schemas/EmailString",
						"description": "Email to allow users to start communicating"
					},
					"relatedSubjects": {
						"items": {
							"$ref": "#/components/schemas/UserSubjectData"
						},
						"type": "array",
						"description": "Links to subjects a user is interested in learning/teaching (depends on type of user)\r\nand user specific information about the user relating to a specific subject"
					},
					"relatedTutors": {
						"items": {
							"$ref": "#/components/schemas/LinkedTutorData"
						},
						"type": "array",
						"description": "Data about the tutors a student is involved with"
					}
				},
				"required": [
					"id",
					"prefferedCountries",
					"prefferedLocales",
					"userName",
					"email",
					"relatedSubjects",
					"relatedTutors"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateStudentResponseBody": {
				"properties": {
					"student": {
						"$ref": "#/components/schemas/PrivateStudentData"
					}
				},
				"required": [
					"student"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_PrivateStudentData.Exclude_keyofPrivateStudentData.id__": {
				"properties": {
					"relatedTutors": {
						"items": {
							"$ref": "#/components/schemas/LinkedTutorData"
						},
						"type": "array",
						"description": "Data about the tutors a student is involved with"
					},
					"email": {
						"$ref": "#/components/schemas/EmailString",
						"description": "Email to allow users to start communicating"
					},
					"relatedSubjects": {
						"items": {
							"$ref": "#/components/schemas/UserSubjectData"
						},
						"type": "array",
						"description": "Links to subjects a user is interested in learning/teaching (depends on type of user)\r\nand user specific information about the user relating to a specific subject"
					},
					"imageUrl": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UrlString"
							}
						],
						"nullable": true,
						"description": "Link to profile image, if defined"
					},
					"introduction": {
						"type": "string",
						"description": "Optional text for users to introduce themselves,\r\ndescribe their experience, what subjects they are interested in etc"
					},
					"prefferedCountries": {
						"items": {
							"$ref": "#/components/schemas/Country"
						},
						"type": "array",
						"description": "The countries the user prefers for content"
					},
					"prefferedLocales": {
						"items": {
							"$ref": "#/components/schemas/LocaleCode"
						},
						"type": "array",
						"description": "The locales a user prefers for content"
					},
					"userName": {
						"type": "string",
						"description": "Username which will be displayed for a user"
					}
				},
				"required": [
					"relatedTutors",
					"email",
					"relatedSubjects",
					"prefferedCountries",
					"prefferedLocales",
					"userName"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_PrivateStudentData.id_": {
				"$ref": "#/components/schemas/Pick_PrivateStudentData.Exclude_keyofPrivateStudentData.id__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"Partial_CreateStudentRequestBody_": {
				"properties": {
					"student": {
						"$ref": "#/components/schemas/Omit_PrivateStudentData.id_"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"GetStudentResponseBody": {
				"properties": {
					"student": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/PrivateStudentData"
							},
							{
								"$ref": "#/components/schemas/PublicStudentData"
							}
						]
					}
				},
				"required": [
					"student"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_GetStudentRequestBody_": {
				"properties": {
					"id": {
						"$ref": "#/components/schemas/StudentId"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"GetStudentsBySubjectsResponseBody": {
				"properties": {
					"students": {
						"items": {
							"$ref": "#/components/schemas/PublicStudentData"
						},
						"type": "array"
					}
				},
				"required": [
					"students"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_GetStudentsBySubjectsRequestBody_": {
				"properties": {
					"subjectIds": {
						"items": {
							"$ref": "#/components/schemas/SubjectId"
						},
						"type": "array"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"UpdateStudentResponseBody": {
				"properties": {
					"result": {
						"$ref": "#/components/schemas/PrivateStudentData"
					}
				},
				"required": [
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_Omit_PrivateStudentData.id__": {
				"properties": {
					"relatedTutors": {
						"items": {
							"$ref": "#/components/schemas/LinkedTutorData"
						},
						"type": "array",
						"description": "Data about the tutors a student is involved with"
					},
					"email": {
						"$ref": "#/components/schemas/EmailString",
						"description": "Email to allow users to start communicating"
					},
					"relatedSubjects": {
						"items": {
							"$ref": "#/components/schemas/UserSubjectData"
						},
						"type": "array",
						"description": "Links to subjects a user is interested in learning/teaching (depends on type of user)\r\nand user specific information about the user relating to a specific subject"
					},
					"imageUrl": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UrlString"
							}
						],
						"nullable": true,
						"description": "Link to profile image, if defined"
					},
					"introduction": {
						"type": "string",
						"description": "Optional text for users to introduce themselves,\r\ndescribe their experience, what subjects they are interested in etc"
					},
					"prefferedCountries": {
						"items": {
							"$ref": "#/components/schemas/Country"
						},
						"type": "array",
						"description": "The countries the user prefers for content"
					},
					"prefferedLocales": {
						"items": {
							"$ref": "#/components/schemas/LocaleCode"
						},
						"type": "array",
						"description": "The locales a user prefers for content"
					},
					"userName": {
						"type": "string",
						"description": "Username which will be displayed for a user"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"Partial_UpdateStudentRequestBody_": {
				"properties": {
					"updates": {
						"$ref": "#/components/schemas/Partial_Omit_PrivateStudentData.id__"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "@adopt-a-student/api",
		"version": "0.0.0",
		"description": "API for adopt-a-student",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Elias Mangoro"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/createStudent": {
			"post": {
				"operationId": "CreateStudent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateStudentResponseBody"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_CreateStudentRequestBody_"
							}
						}
					}
				}
			}
		},
		"/getStudent": {
			"post": {
				"operationId": "GetStudent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetStudentResponseBody"
								}
							}
						}
					}
				},
				"description": "Retreives data about a student user. If the student user owns the data then they get all the data, otherwise it is restricted to 'public' data.",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_GetStudentRequestBody_"
							}
						}
					}
				}
			}
		},
		"/getStudentsBySubjects": {
			"post": {
				"operationId": "GetStudentsBySubjects",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetStudentsBySubjectsResponseBody"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_GetStudentsBySubjectsRequestBody_"
							}
						}
					}
				}
			}
		},
		"/updateStudent": {
			"post": {
				"operationId": "UpdateStudent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UpdateStudentResponseBody"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_UpdateStudentRequestBody_"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}